rgw multisite: use get_config_master_client for radosgw_admin task
