+++ qa/suites/rgw/singleton/all/radosgw-admin-data-sync.yaml
@@ -63,4 +63,3 @@ tasks:
 - sleep:
     duration: 30
 - radosgw-admin:
-    master_client: client.0
