+++ qa/suites/rgw/singleton/all/radosgw-admin-multi-region.yaml
@@ -65,4 +65,3 @@ tasks:
       dest: client.1
       metadata-only: true
 - radosgw-admin:
-    master_client: client.0
