+++ qa/suites/rgw/singleton/all/radosgw-admin.yaml
@@ -18,4 +18,3 @@ tasks:
 - rgw:
     client.0:
 - radosgw-admin:
-    master_client: client.0
